construct tower, upgrade tower, sell tower, wave fix, consume gold, campaign mode, return level selection after victory or defeat...

+++ tower-defense/tower-defense/src/com/git/tdgame/gameActor/tower/Tower.java
@@ -39,7 +39,7 @@ public class Tower extends Actor
     	//properties
     	this.width = Integer.valueOf(properties.get("width"));
     	this.height = Integer.valueOf(properties.get("height"));
-    	this.fireRate = Integer.valueOf(properties.get("fireRate"));
+    	this.fireRate = Float.valueOf(properties.get("fireRate"));
     	this.range = Integer.valueOf(properties.get("range"));
     	this.cost = Integer.valueOf(properties.get("cost"));
     	this.projectileModel = gson.fromJson(properties.get("projectile"), ProjectileModel.class);
